Changed places store -> users
