+++ store/serializers.py
@@ -1,6 +1,4 @@
 from rest_framework import serializers
-from rest_framework.validators import UniqueValidator
-from django.contrib.auth.models import User
 from store.models import Product, Category
 
 
@@ -19,18 +17,3 @@ class ProductSerializer(serializers.ModelSerializer):
         }
 
 
-class SignupSerializer(serializers.ModelSerializer):
-    email = serializers.EmailField(validators=[UniqueValidator(queryset=User.objects.all())])
-    password = serializers.CharField(max_length=128, write_only=True)
-
-    class Meta:
-        model = User
-        fields = ('id', 'username', 'email', 'password')
-
-    def create(self, validated_data):
-        user = User.objects.create_user(
-            username=validated_data["username"],
-            email=validated_data['email'],
-            password=validated_data['password']
-        )
-        return user
