+++ users/serializers.py
@@ -1,6 +1,7 @@
 from rest_framework import serializers
 from django.contrib.auth.models import User
 from users.models import Profile
+from rest_framework.validators import UniqueValidator
 
 
 class ProfileSerializer(serializers.ModelSerializer):
@@ -15,3 +16,20 @@ class UserProfileSerializers(serializers.ModelSerializer):
     class Meta:
         model = User
         fields = ('id', 'username', 'email', 'first_name', 'last_name', 'profile')
+
+
+class SignupSerializer(serializers.ModelSerializer):
+    email = serializers.EmailField(validators=[UniqueValidator(queryset=User.objects.all())])
+    password = serializers.CharField(max_length=128, write_only=True)
+
+    class Meta:
+        model = User
+        fields = ('id', 'username', 'email', 'password')
+
+    def create(self, validated_data):
+        user = User.objects.create_user(
+            username=validated_data["username"],
+            email=validated_data['email'],
+            password=validated_data['password']
+        )
+        return user
